+++ src/input/r_matroska.h
@@ -190,6 +190,7 @@ struct kax_track_t {
   void handle_packetizer_pixel_dimensions();
   void handle_packetizer_default_duration();
   void handle_packetizer_output_sampling_freq();
+  void handle_packetizer_codec_delay();
   void fix_display_dimension_parameters();
   void add_track_tags_to_identification(mtx::id::info_c &info);
   void discard_track_statistics_tags();